Address bug report #11, by using 64-bit offsets and lengths internally.  Create a regression test to capture this case

+++ LibOS/shim/include/shim_fs.h
@@ -69,7 +69,7 @@ struct shim_fs_ops {
     int (*move) (const char * trim_old_name, const char * trim_new_name);
     int (*copy) (const char * trim_old_name, const char * trim_new_name);
 
-    int (*truncate) (struct shim_handle * hdl, int len);
+    int (*truncate) (struct shim_handle * hdl, uint64_t len);
 
     /* hstat: get status of the file */
     int (*hstat) (struct shim_handle * hdl, struct stat * buf);
